+++ package.json
@@ -11,15 +11,15 @@
   "scripts": {
     "prebuild": "rm -rf ./prod",
     "build": "tsc",
-    "lint": "eslint src --max-warnings 0",
+    "lint": "eslint src test --max-warnings 0",
     "start": "node ./prod/index.js",
-    "test": "ts-mocha -u tdd ./src/test/*.spec.ts"
+    "test": "ts-mocha -u tdd test/**/*.spec.ts --recursive --exit --timeout 10000"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/nhcarrigan/discord-rpc.git"
   },
-  "author": "Nicholas Carrigan",
+  "author": "Naomi Carrigan",
   "license": "AGPL-3.0-or-later",
   "bugs": {
     "url": "https://github.com/nhcarrigan/discord-rpc/issues"
